Refactor Pimoroni PIM447 trackball driver: Add toggle mode and sleep control
This commit adds new functionality to the Pimoroni PIM447 trackball driver. It introduces the ability to toggle between different modes and control the sleep state of the device. The behavior header file include guard is updated to include the new toggle mode and sleep control constants. The device tree and driver compatibility are also updated to support these changes. This refactor improves the overall functionality and flexibility of the PIM447 trackball driver.